+++ changelog.md
@@ -7,4 +7,4 @@
 
 ### v0.2.0
 
-* Extracted Marionette.Async in to it's own repository
+* Extracted Marionette.Async in to its own repository